# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 관리의달인 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /commissions:
    get:
      operationId: listCommissions
      tags: [commissions]
      summary: List commission records
      responses:
        "200":
          description: Commission list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commission"
    post:
      operationId: createCommission
      tags: [commissions]
      summary: Create a commission record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommissionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
  /commissions/{id}:
    patch:
      operationId: updateCommission
      tags: [commissions]
      summary: Update commission status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommissionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
  /commissions/auto-settle:
    post:
      operationId: autoSettleCommission
      tags: [commissions]
      summary: Auto-create commission record when contract completes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoSettleCommissionBody"
      responses:
        "201":
          description: Commission auto-settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoSettleCommissionResponse"

  # ── RFQs (견적요청) ──────────────────────────────────────
...
  /commissions/rates:
    get:
      operationId: listCommissionRates
      tags: [commissions]
      summary: List commission rates by category
      responses:
        "200":
          description: Rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommissionRate"
    put:
      operationId: upsertCommissionRate
      tags: [commissions]
      summary: Upsert commission rate for a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCommissionRateBody"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionRate"
...
  /commissions/{id}/transition:
    post:
      operationId: transitionCommission
      tags: [commissions]
      summary: Transition commission status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionTransitionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
  /commissions/{id}/events:
    get:
      operationId: listCommissionEvents
      tags: [commissions]
      summary: List commission status event history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommissionEvent"
components:
  schemas:
...
    Commission:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        contractAmount:
          type: number
        commissionRate:
          type: number
        commissionAmount:
          type: number
        status:
          type: string
          enum: [pending, billed, collected, completed, cancelled, confirmed, paid]
        matchedDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        rfqId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        billedAt:
          type: ["string", "null"]
          format: date-time
        collectedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        invoiceNumber:
          type: ["string", "null"]
        invoiceIssuedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - vendorId
        - vendorName
        - contractAmount
        - commissionRate
        - commissionAmount
        - status
        - matchedDate
        - createdAt
        - updatedAt
    CreateCommissionBody:
      type: object
      properties:
        vendorId:
          type: integer
        vendorName:
          type: string
        contractAmount:
          type: number
        commissionRate:
          type: number
        commissionAmount:
          type: number
        matchedDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
      required:
        - vendorId
        - vendorName
        - contractAmount
        - commissionRate
        - commissionAmount
        - matchedDate
    UpdateCommissionBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, billed, collected, completed, cancelled, confirmed, paid]
        notes:
          type: ["string", "null"]
...
    AutoSettleCommissionBody:
      type: object
      properties:
        vendorId:
          type: integer
        vendorName:
          type: string
        contractAmount:
          type: number
        commissionRate:
          type: number
        inspectionId:
          type: ["integer", "null"]
        rfqId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - vendorId
        - vendorName
        - contractAmount
    AutoSettleCommissionResponse:
      type: object
      properties:
        commission:
          $ref: "#/components/schemas/Commission"
        message:
          type: string
      required:
        - commission
        - message
...
    CommissionRate:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        rateType:
          type: string
          enum: [fixed, sliding]
        fixedRate:
          type: number
        slidingRules:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - rateType
        - fixedRate
    SlidingRule:
      type: object
      properties:
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        ratePercent:
          type: number
      required:
        - minAmount
        - ratePercent
    UpsertCommissionRateBody:
      type: object
      properties:
        category:
          type: string
        rateType:
          type: string
          enum: [fixed, sliding]
        fixedRate:
          type: number
        slidingRules:
          type: array
          items:
            $ref: "#/components/schemas/SlidingRule"
        description:
          type: ["string", "null"]
      required:
        - category
        - rateType
...
    CommissionTransitionBody:
      type: object
      properties:
        toStatus:
          type: string
          enum: [pending, billed, collected, completed, cancelled]
        reason:
          type: ["string", "null"]
      required:
        - toStatus
    CommissionEvent:
      type: object
      properties:
        id:
          type: integer
        commissionId:
          type: integer
        fromStatus:
          type: ["string", "null"]
        toStatus:
          type: string
        reason:
          type: ["string", "null"]
        actorId:
          type: ["integer", "null"]
        actorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - commissionId
        - toStatus
        - createdAt