# lib/api-zod/src/generated/types/commissionRate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 관리의달인 API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommissionRateRateType } from "./commissionRateRateType";

export interface CommissionRate {
  id: number;
  category: string;
  rateType: CommissionRateRateType;
  fixedRate: number;
  /** @nullable */
  slidingRules?: string | null;
  /** @nullable */
  description?: string | null;
  updatedAt?: Date;
}