# lib/api-zod/src/generated/types/slidingRule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 관리의달인 API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SlidingRule {
  minAmount: number;
  /** @nullable */
  maxAmount?: number | null;
  ratePercent: number;
}