# lib/api-zod/src/generated/types/upsertCommissionRateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 관리의달인 API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SlidingRule } from "./slidingRule";
import type { UpsertCommissionRateBodyRateType } from "./upsertCommissionRateBodyRateType";

export interface UpsertCommissionRateBody {
  category: string;
  rateType: UpsertCommissionRateBodyRateType;
  fixedRate?: number;
  slidingRules?: SlidingRule[];
  /** @nullable */
  description?: string | null;
}